+++ src/commands/bub.js
@@ -3,7 +3,7 @@ const fs = require("fs");
 const chalk = require("chalk");
 
 // Get functions
-const _replaceSpaces = require("../functions/replaceSpaces");
+const _parseDoubleQuotes = require("../functions/parseQuotes");
 const _convertAbsolute = require("../functions/convAbs");
 const _fatalError = require("../functions/fatalError");
 
@@ -99,7 +99,7 @@ const _interpretBubbleFile = (
 const bub = (intCmds, file, ...args) => {
   try {
     // Replace spaces and convert to an absolute path
-    file = _convertAbsolute(_replaceSpaces(file));
+    file = _convertAbsolute(_parseDoubleQuotes([file, ...args])[0]);
 
     // Initialize checker
     const fileChk = new Checks(file);