+++ src/commands/mkfile.js
@@ -4,7 +4,7 @@ const chalk = require("chalk");
 const { question } = require("readline-sync");
 
 // Get functions
-const _replaceSpaces = require("../functions/replaceSpaces");
+const _parseDoubleQuotes = require("../functions/parseQuotes");
 const _convertAbsolute = require("../functions/convAbs");
 const _fatalError = require("../functions/fatalError");
 
@@ -43,7 +43,7 @@ const Checks = require("../classes/Checks");
 const mkfile = (file, ...args) => {
   try {
     // Replace spaces and then convert the file to an absolute path
-    file = _convertAbsolute(_replaceSpaces(file));
+    file = _convertAbsolute(_parseDoubleQuotes([file, ...args])[0]);
 
     // Initialize checker
     const fileChk = new Checks(file);